chore: Rename orders page for module consistency
