+++ apps/dashboard/src/pages/_sites/[site]/orders/index.tsx
@@ -1,1 +1,1 @@
-export { default, getServerSideProps } from "@/modules/Orders/orders";
+export { default, getServerSideProps } from "@/modules/Orders";
