+++ apps/dashboard/src/pages/orders/index.tsx
@@ -1,1 +1,1 @@
-export { default, getServerSideProps } from "@/modules/Orders/orders";
+export { default, getServerSideProps } from "@/modules/Orders";
